Simplify code: remove key on top-level <Outlet /> and use <Outlet /> in search page
