+++ client/src/routes/search/empty.tsx
@@ -1,30 +1,29 @@
-import React, { Suspense, useDeferredValue } from 'react';
-import { useParams } from 'react-router-dom';
+import React, { useDeferredValue } from 'react';
+import {
+  useNavigate,
+  Outlet,
+  useParams,
+  useOutletContext,
+} from 'react-router-dom';
 import { SearchIcon } from 'lucide-react';
-import Skeleton from '../../components/Skeleton';
 import SearchQueryForm from '../../components/SearchQueryForm';
 import Page from '../../components/Page';
-import { RouteComponent as QueryRouteComponent } from './query';
 
-export const LoadingState = () => {
-  return (
-    <Page className="p-[var(--main-content--padding)]">
-      <div className="flex flex-col gap-4">
-        <Skeleton.TileGrid
-          gap="1rem"
-          template={<Skeleton.MediaTile description coverPhotoShape="circle" />}
-          tileCount={15}
-          minTileWidth="200px"
-        />
-      </div>
-    </Page>
-  );
-};
+type ContextType = { query: string };
+
+export function useSearchQuery() {
+  return useOutletContext<ContextType>();
+}
 
 export const RouteComponent = () => {
   const params = useParams();
+  const navigate = useNavigate();
   const [query, setQuery] = React.useState(params.query ?? '');
   const deferredQuery = useDeferredValue(query);
+
+  React.useEffect(() => {
+    navigate('/search/' + encodeURIComponent(deferredQuery));
+  }, [deferredQuery, navigate]);
 
   return (
     <Page className="p-[var(--main-content--padding)]">
@@ -33,9 +32,7 @@ export const RouteComponent = () => {
         onSubmit={(data) => setQuery(data.query)}
       />
       {deferredQuery ? (
-        <Suspense fallback={<LoadingState />}>
-          <QueryRouteComponent query={deferredQuery} />
-        </Suspense>
+        <Outlet context={{ query: deferredQuery } satisfies ContextType} />
       ) : (
         <Page.EmptyState
           icon={<SearchIcon />}
